requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
ultralytics==8.1.0
opencv-python==4.9.0.80
numpy==1.26.3
aiofiles==23.2.1
